+++ src/main/java/miyucomics/hexical/casting/patterns/conjure/OpSnowball.kt
@@ -8,23 +8,24 @@ import at.petrak.hexcasting.api.casting.getVec3
 import at.petrak.hexcasting.api.casting.iota.EntityIota
 import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.misc.MediaConstants
-import net.minecraft.entity.projectile.FireballEntity
+import net.minecraft.entity.projectile.thrown.SnowballEntity
 import net.minecraft.util.math.Vec3d
 
-class OpGhastFireball : SpellAction {
+class OpSnowball : SpellAction {
 	override val argc = 1
 	override fun execute(args: List<Iota>, env: CastingEnvironment): SpellAction.Result {
 		val position = args.getVec3(0, argc)
 		env.assertVecInRange(position)
-		return SpellAction.Result(Spell(position), MediaConstants.DUST_UNIT * 3, listOf())
+		return SpellAction.Result(Spell(position), MediaConstants.DUST_UNIT / 2, listOf())
 	}
 
 	private data class Spell(val position: Vec3d) : RenderedSpell {
 		override fun cast(env: CastingEnvironment) {}
 		override fun cast(env: CastingEnvironment, image: CastingImage): CastingImage {
-			val fireball = FireballEntity(env.world, env.castingEntity, position.x, position.y, position.z, 1)
-			env.world.spawnEntity(fireball)
-			return image.copy(stack = image.stack.toList().plus(EntityIota(fireball)))
+			val snowball = SnowballEntity(env.world, position.x, position.y, position.z)
+			snowball.owner = env.castingEntity
+			env.world.spawnEntity(snowball)
+			return image.copy(stack = image.stack.toList().plus(EntityIota(snowball)))
 		}
 	}
 }